+++ src/bl64-bsh-env.bash
@@ -2,10 +2,10 @@
 # BashLib64 / Module / Globals / Interact with Bash shell
 #######################################
 
-export BL64_BSH_VERSION='2.0.1'
+export BL64_BSH_VERSION='2.1.0'
 
 export BL64_BSH_MODULE='0'
 
 export BL64_BSH_VERSION_BASH=''
 
-export _BL64_BSH_TXT_UNSUPPORTED='BashLib64 is not supported in the current Bash version'
+export _BL64_BSH_TXT_IMPORT_YAML='convert and import shell variables from YAML file'