+++ src/bl64-bsh-setup.bash
@@ -24,11 +24,14 @@ function bl64_bsh_setup() {
   bl64_lib_module_imported 'BL64_DBG_MODULE' &&
     bl64_lib_module_imported 'BL64_CHECK_MODULE' &&
     bl64_lib_module_imported 'BL64_FMT_MODULE' &&
+    bl64_lib_module_imported 'BL64_XSV_MODULE' &&
+    bl64_lib_module_imported 'BL64_TXT_MODULE' &&
+    bl64_lib_module_imported 'BL64_FS_MODULE' &&
     _bl64_bsh_set_version &&
     BL64_BSH_MODULE="$BL64_VAR_ON"
 
   bl64_dbg_lib_show_function
-  bl64_check_alert_module_setup ' bsh'
+  bl64_check_alert_module_setup 'bsh'
 }
 
 #######################################